+++ Editor/LLMClientEditor.cs
@@ -3,7 +3,6 @@ using UnityEditor;
 using UnityEngine;
 using System.Reflection;
 using System.Collections.Generic;
-using System.Linq;
 
 namespace LLMUnity
 {
@@ -13,17 +12,21 @@ namespace LLMUnity
         protected int buttonWidth = 150;
         Type[] orderedTypes = new Type[] { typeof(LLM), typeof(LLMClient) };
 
+        public void Space()
+        {
+            EditorGUILayout.Space((int)EditorGUIUtility.singleLineHeight / 2);
+        }
+
         public void AddScript(SerializedObject llmScriptSO)
         {
             var scriptProp = llmScriptSO.FindProperty("m_Script");
             EditorGUILayout.PropertyField(scriptProp);
-            EditorGUILayout.Space((int)EditorGUIUtility.singleLineHeight / 2);
         }
 
-        public void AddAdvancedOptionsToggle(SerializedObject llmScriptSO)
+        public void AddOptionsToggle(SerializedObject llmScriptSO, string propertyName, string name)
         {
-            SerializedProperty advancedOptionsProp = llmScriptSO.FindProperty("advancedOptions");
-            string toggleText = (advancedOptionsProp.boolValue ? "Hide" : "Show") + " Advanced Options";
+            SerializedProperty advancedOptionsProp = llmScriptSO.FindProperty(propertyName);
+            string toggleText = (advancedOptionsProp.boolValue ? "Hide" : "Show") + " " + name;
             GUIStyle style = new GUIStyle("Button");
             if (advancedOptionsProp.boolValue)
                 style.normal = new GUIStyleState() { background = Texture2D.grayTexture };
@@ -31,22 +34,70 @@ namespace LLMUnity
             {
                 advancedOptionsProp.boolValue = !advancedOptionsProp.boolValue;
             }
-            EditorGUILayout.Space();
+        }
+
+        public void AddOptionsToggles(SerializedObject llmScriptSO)
+        {
+            EditorGUILayout.BeginHorizontal();
+            AddOptionsToggle(llmScriptSO, "advancedOptions", "Advanced Options");
+            AddOptionsToggle(llmScriptSO, "expertOptions", "Expert Options");
+            EditorGUILayout.EndHorizontal();
+            Space();
         }
 
         public void AddServerSettings(SerializedObject llmScriptSO)
         {
-            ShowPropertiesOfClass("Server Settings", llmScriptSO, orderedTypes, typeof(ServerAttribute), typeof(ServerAdvancedAttribute), true);
+            List<Type> attributeClasses = new List<Type> { typeof(ServerAttribute) };
+            if (llmScriptSO.FindProperty("advancedOptions").boolValue) attributeClasses.Add(typeof(ServerAdvancedAttribute));
+            ShowPropertiesOfClass("Server Settings", llmScriptSO, orderedTypes, attributeClasses, true);
         }
 
-        public void AddModelSettings(SerializedObject llmScriptSO, bool showHeader = true)
+        public void AddModelAddonLoaders(SerializedObject llmScriptSO, LLMClient llmScript, bool layout = true)
         {
-            ShowPropertiesOfClass(showHeader ? "Model Settings" : "", llmScriptSO, orderedTypes, typeof(ModelAttribute), typeof(ModelAdvancedAttribute), true);
+            if (llmScriptSO.FindProperty("advancedOptions").boolValue)
+            {
+                if (layout) EditorGUILayout.BeginHorizontal();
+                if (GUILayout.Button("Load grammar", GUILayout.Width(buttonWidth)))
+                {
+                    EditorApplication.delayCall += () =>
+                    {
+                        string path = EditorUtility.OpenFilePanelWithFilters("Select a gbnf grammar file", "", new string[] { "Grammar Files", "gbnf" });
+                        if (!string.IsNullOrEmpty(path))
+                        {
+                            llmScript.SetGrammar(path);
+                        }
+                    };
+                }
+                if (layout) EditorGUILayout.EndHorizontal();
+            }
+        }
+
+        public void AddModelSettings(SerializedObject llmScriptSO)
+        {
+            List<Type> attributeClasses = new List<Type> { typeof(ModelAttribute), typeof(ModelAddonAttribute) };
+            if (llmScriptSO.FindProperty("advancedOptions").boolValue)
+            {
+                attributeClasses.Add(typeof(ModelAddonAdvancedAttribute));
+                attributeClasses.Add(typeof(ModelAdvancedAttribute));
+            }
+            if (llmScriptSO.FindProperty("expertOptions").boolValue)
+            {
+                attributeClasses.Add(typeof(ModelExpertAttribute));
+            }
+            ShowPropertiesOfClass("", llmScriptSO, orderedTypes, attributeClasses, true);
+        }
+
+        public void AddModelLoadersSettings(SerializedObject llmScriptSO, LLMClient llmScript)
+        {
+            if (!llmScriptSO.FindProperty("advancedOptions").boolValue) return; // at the moment we only have advanced parameters here
+            EditorGUILayout.LabelField("Model Settings", EditorStyles.boldLabel);
+            AddModelAddonLoaders(llmScriptSO, llmScript);
+            AddModelSettings(llmScriptSO);
         }
 
         public void AddChatSettings(SerializedObject llmScriptSO)
         {
-            ShowPropertiesOfClass("Chat Settings", llmScriptSO, orderedTypes, typeof(ChatAttribute), null, false);
+            ShowPropertiesOfClass("Chat Settings", llmScriptSO, orderedTypes, new List<Type> { typeof(ChatAttribute) }, false);
         }
 
         public override void OnInspectorGUI()
@@ -59,9 +110,9 @@ namespace LLMUnity
             AddScript(llmScriptSO);
             GUI.enabled = true;
             EditorGUI.BeginChangeCheck();
-            AddAdvancedOptionsToggle(llmScriptSO);
+            AddOptionsToggles(llmScriptSO);
             AddServerSettings(llmScriptSO);
-            AddModelSettings(llmScriptSO);
+            AddModelLoadersSettings(llmScriptSO, llmScript);
             AddChatSettings(llmScriptSO);
             EditorGUI.EndChangeCheck();
             if (EditorGUI.EndChangeCheck())
@@ -70,13 +121,10 @@ namespace LLMUnity
             llmScriptSO.ApplyModifiedProperties();
         }
 
-        public List<SerializedProperty> GetPropertiesOfClass(SerializedObject so, Type[] targetClasses, Type attributeClass = null, Type attributeAdvancedClass = null)
+        public List<SerializedProperty> GetPropertiesOfClass(SerializedObject so, Type[] targetClasses, List<Type> attributeClasses)
         {
             // display a property if it belongs to a certain class and/or has a specific attribute class
             List<SerializedProperty> properties = new List<SerializedProperty>();
-            List<Type> attributeClasses = new List<Type> { attributeClass };
-            if (so.FindProperty("advancedOptions").boolValue)
-                attributeClasses.Add(attributeAdvancedClass);
             foreach (Type attrClass in attributeClasses)
             {
                 if (attrClass == null) continue;
@@ -97,14 +145,30 @@ namespace LLMUnity
             return properties;
         }
 
-        public void ShowPropertiesOfClass(string title, SerializedObject so, Type[] targetClasses, Type attributeClass = null, Type attributeAdvancedClass = null, bool addSpace = true)
+        public void ShowPropertiesOfClass(string title, SerializedObject so, Type[] targetClasses, List<Type> attributeClasses, bool addSpace = true)
         {
             // display a property if it belongs to a certain class and/or has a specific attribute class
-            List<SerializedProperty> properties = GetPropertiesOfClass(so, targetClasses, attributeClass, attributeAdvancedClass);
+            List<SerializedProperty> properties = GetPropertiesOfClass(so, targetClasses, attributeClasses);
             if (properties.Count == 0) return;
             if (title != "") EditorGUILayout.LabelField(title, EditorStyles.boldLabel);
-            foreach (SerializedProperty prop in properties) EditorGUILayout.PropertyField(prop);
-            if (addSpace) EditorGUILayout.Space(EditorGUIUtility.singleLineHeight);
+            foreach (SerializedProperty prop in properties)
+            {
+                Attribute floatAttr = GetPropertyAttribute(prop, typeof(FloatAttribute));
+                Attribute intAttr = GetPropertyAttribute(prop, typeof(IntAttribute));
+                if (floatAttr != null)
+                {
+                    EditorGUILayout.Slider(prop, ((FloatAttribute)floatAttr).Min, ((FloatAttribute)floatAttr).Max, new GUIContent(prop.displayName));
+                }
+                else if (intAttr != null)
+                {
+                    EditorGUILayout.IntSlider(prop, ((IntAttribute)intAttr).Min, ((IntAttribute)intAttr).Max, new GUIContent(prop.displayName));
+                }
+                else
+                {
+                    EditorGUILayout.PropertyField(prop);
+                }
+            }
+            if (addSpace) Space();
         }
 
         public bool PropertyInClass(SerializedProperty prop, Type targetClass, Type attributeClass = null)
@@ -112,10 +176,10 @@ namespace LLMUnity
             // check if a property belongs to a certain class and/or has a specific attribute class
             FieldInfo field = prop.serializedObject.targetObject.GetType().GetField(prop.name,
                 BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic);
-            return field != null && field.DeclaringType == targetClass && (attributeClass == null || AttributeInProperty(prop, attributeClass));
+            return field != null && field.DeclaringType == targetClass && (attributeClass == null || GetPropertyAttribute(prop, attributeClass) != null);
         }
 
-        public bool AttributeInProperty(SerializedProperty prop, Type attributeClass)
+        public Attribute GetPropertyAttribute(SerializedProperty prop, Type attributeClass)
         {
             // check if a property has a specific attribute class
             foreach (var pathSegment in prop.propertyPath.Split('.'))
@@ -129,13 +193,13 @@ namespace LLMUnity
                         foreach (Attribute attr in fieldInfo.GetCustomAttributes(attributeClass, true))
                         {
                             if (attr.GetType() == attributeClass)
-                                return true;
+                                return attr;
                         }
                     }
                     targetType = targetType.BaseType;
                 }
             }
-            return false;
+            return null;
         }
     }
 }